ntdll: Use UINT instead of ULONG in ioctl helpers.
To avoid typecasts in debug traces.

+++ dlls/ntdll/unix/cdrom.c
@@ -2814,8 +2814,8 @@ static NTSTATUS GetInquiryData(int fd, PSCSI_ADAPTER_BUS_INFO BufferOut, DWORD O
  *		cdrom_DeviceIoControl
  */
 NTSTATUS cdrom_DeviceIoControl( HANDLE device, HANDLE event, PIO_APC_ROUTINE apc, void *apc_user,
-                                IO_STATUS_BLOCK *io, ULONG code, void *in_buffer,
-                                ULONG in_size, void *out_buffer, ULONG out_size )
+                                IO_STATUS_BLOCK *io, UINT code, void *in_buffer,
+                                UINT in_size, void *out_buffer, UINT out_size )
 {
     DWORD       sz = 0;
     NTSTATUS    status = STATUS_SUCCESS;
